refactor(editions): move editions loader from app/lib to utils
editions.ts is a plain data util rather than page-level code, so it
belongs alongside the other shared utils. Pure move — updates the 13
import sites, no behaviour change.
