+++ app/api/fighters/injuries/route.ts
@@ -1,6 +1,6 @@
 import { createClient } from "@/utils/supabase/server";
 import { NextResponse } from 'next/server';
-import { getEditionIdBySlug } from '@/app/lib/editions';
+import { getEditionIdBySlug } from '@/utils/editions';
 
 /**
  * Lists the lasting injury (or rig glitch) catalog.
